Fixed several critical autocycling bugs (related Bug-Tracker-IDs: 3096271, 3098180). Fixed a bug causing an unhandled exception when a update server was unreachable. Fixed a bug causing to show error dialogs on update checking during application startup. Optimized wallpaper cycling methods.
git-svn-id: svn+ssh://svn.code.sf.net/p/wallpaperman/code@8 1d17e2ee-c682-4ae3-af1b-aef583486f15

+++ Wallpaper Manager/Data Layer/GeneralConfig.cs
@@ -192,7 +192,7 @@ namespace WallpaperManager.Data {
     public Byte LastActiveListSize {
       get { return this.lastActiveListSize; }
       set {
-        if (!value.IsBetween((Byte)1, WallpaperChanger.LastActiveListSizeMax)) {
+        if (!value.IsBetween(1, WallpaperChanger.LastActiveListSizeMax)) {
           throw new ArgumentOutOfRangeException(ExceptionMessages.GetValueOutOfRange(
             null, value, 
             1.ToString(CultureInfo.CurrentCulture),
